Color Coherence Vector
We are going to talk about one of the commonly known techniques used in image retrieval, which is the Color coherence vector. It is an image descriptor (or more specifically, it is a color descriptor) that extracts color-related features from the image, which can be used as a low-dimensional representation of this image.